Issue #3661 - move WebSocketSessionListener to WebSocket-API
