+++ jetty-websocket/jetty-websocket-common/src/main/java/org/eclipse/jetty/websocket/common/SessionTracker.java
@@ -25,6 +25,7 @@ import java.util.concurrent.CopyOnWriteArrayList;
 import org.eclipse.jetty.util.component.AbstractLifeCycle;
 import org.eclipse.jetty.util.component.LifeCycle;
 import org.eclipse.jetty.websocket.api.Session;
+import org.eclipse.jetty.websocket.api.WebSocketSessionListener;
 
 public class SessionTracker extends AbstractLifeCycle implements WebSocketSessionListener
 {
@@ -36,13 +37,13 @@ public class SessionTracker extends AbstractLifeCycle implements WebSocketSessio
     }
 
     @Override
-    public void onWebSocketSessionOpened(WebSocketSession session)
+    public void onWebSocketSessionOpened(Session session)
     {
         sessions.add(session);
     }
 
     @Override
-    public void onWebSocketSessionClosed(WebSocketSession session)
+    public void onWebSocketSessionClosed(Session session)
     {
         sessions.remove(session);
     }
