+++ jetty-websocket/jetty-websocket-common/src/main/java/org/eclipse/jetty/websocket/common/WebSocketContainer.java
@@ -23,9 +23,10 @@ import java.util.concurrent.Executor;
 import java.util.function.Consumer;
 
 import org.eclipse.jetty.websocket.api.Session;
+import org.eclipse.jetty.websocket.api.WebSocketSessionListener;
 
 /**
- * Generic interface to the Container (server or client) that jetty-websocket-common can use
+ * Generic interface to the Container (server or client)
  */
 public interface WebSocketContainer
 {
