+++ jetty-websocket/jetty-websocket-tests/src/test/java/org/eclipse/jetty/websocket/tests/client/ClientOpenSessionTracker.java
@@ -21,9 +21,9 @@ package org.eclipse.jetty.websocket.tests.client;
 import java.util.concurrent.CountDownLatch;
 
 import org.eclipse.jetty.io.Connection;
+import org.eclipse.jetty.websocket.api.Session;
+import org.eclipse.jetty.websocket.api.WebSocketSessionListener;
 import org.eclipse.jetty.websocket.client.WebSocketClient;
-import org.eclipse.jetty.websocket.common.WebSocketSession;
-import org.eclipse.jetty.websocket.common.WebSocketSessionListener;
 
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -64,12 +64,12 @@ public class ClientOpenSessionTracker implements Connection.Listener, WebSocketS
     }
 
     @Override
-    public void onWebSocketSessionOpened(WebSocketSession session)
+    public void onWebSocketSessionOpened(Session session)
     {
     }
 
     @Override
-    public void onWebSocketSessionClosed(WebSocketSession session)
+    public void onWebSocketSessionClosed(Session session)
     {
         this.closeSessionLatch.countDown();
     }
